Update README.md
Sick Days

Write a program that prompts the user for the number of tellers at Nation’s Bank in Hyattsville. For each worker the program should ask for the number of days that worker was out sick for the last year. The output should provide the number of tellers and the total number of days missed by all the tellers.

See the sample output below.

How many tellers worked at Nation's Bank last year?
3
How many sick days was teller 1 out for, during the last year ?
2
How many sick days was teller 2 out for, during the last year ?
4
How many sick days was teller 3 out for, during the last year ?
7
The 3 tellers were out sick for a total of 13 days during the last year.